+++ changelog.md
@@ -9,6 +9,7 @@
 ### Refactor
 
 * [1783](https://github.com/zeta-chain/node/pull/1783) - refactor zetaclient metrics naming and structure
+* [1774](https://github.com/zeta-chain/node/pull/1774) - split params and config in zetaclient
 
 ### Tests